+++ internal/filepath/glob.go
@@ -1,6 +1,7 @@
 package filepath
 
 import (
+	"os"
 	"path/filepath"
 	"runtime"
 	"strings"
@@ -22,11 +23,17 @@ func hasMeta(path string) bool {
 // a list of explicit file paths. The paths are de-duplicated but are not
 // sorted.
 func Globs(paths []string) ([]string, error) {
-	expandedPaths := make([]string, 0, len(paths))
-	seenPaths := make(map[string]struct{}, len(paths))
+	var expandedPaths []string
+	seenPaths := map[string]struct{}{}
 
 	for _, path := range paths {
-		globbed, err := filepath.Glob(path)
+		var globbed []string
+		var err error
+		if segments := strings.Split(path, "**"); len(segments) == 1 {
+			globbed, err = filepath.Glob(path)
+		} else {
+			globbed, err = superGlobs(segments)
+		}
 		if err != nil {
 			return nil, err
 		}
@@ -45,5 +52,43 @@ func Globs(paths []string) ([]string, error) {
 	}
 
 	return expandedPaths, nil
+}
 
+// Inspired by https://github.com/yargevad/filepathx/blob/master/filepathx.go
+func superGlobs(segments []string) ([]string, error) {
+	matches := map[string]struct{}{"": {}}
+
+	for i, segment := range segments {
+		newMatches := map[string]struct{}{}
+		lastSegment := (len(segments) - 1) == i
+
+		for match := range matches {
+			paths, err := filepath.Glob(match + segment)
+			if err != nil {
+				return nil, err
+			}
+			for _, path := range paths {
+				if err := filepath.Walk(path, func(newPath string, info os.FileInfo, err error) error {
+					if err != nil {
+						return err
+					}
+					if lastSegment && info.IsDir() {
+						return nil
+					}
+					newMatches[newPath] = struct{}{}
+					return nil
+				}); err != nil {
+					return nil, err
+				}
+			}
+		}
+
+		matches = newMatches
+	}
+
+	matchSlice := make([]string, 0, len(matches))
+	for path := range matches {
+		matchSlice = append(matchSlice, path)
+	}
+	return matchSlice, nil
 }